Updated parse.php to include position, placement, and site into listings table. Also changed file to accept url variable instead of code change for testing purposes
Created buildnav.php to create a position table and fill in the position, placement, and site id with a count of the current ads that are online. It will delete any positions that end up with 0 ads for a particular site.

Created delete.php to remove all old ads since it will need to be run seperate from parse.php (which will only be run as new ads come in).

It is important to note delete.php should be run prior to buildnav.php first thing in the morning. Also buildnav.php should be run immediately after parse.php is finished running to provide an accurate ad number listing.